Refactor : removed useless functions from config.py, deleted heroic_asset.py and drag_drop_file.py. config.py has been reworked and might need a rename for user_config.py
