+++ src/core/scanner.py
@@ -8,9 +8,7 @@ import yaml
 from gi.repository import GLib
 
 from core.config import update_user_config, write_yaml, load_yaml
-from core.heroic_asset import download_heroic_assets
 from typing import List, Dict, Optional, Any
-
 
 def slugify(text: str) -> str:
     return re.sub(r'[^a-z0-9]', '', text.lower())
@@ -57,7 +55,7 @@ def get_heroic_library_paths() -> Dict[str, Optional[str]]:
 
     return paths
 
-# Image type 1 = game image ; Image type 2 = banner
+# Now returns a dictionary so you don't need two methods to get 2 images
 def find_game_art(app_id: str | int, platform: str, steam_base: Optional[str]) -> dict:
     art = {"hero": None, "poster": None}
     if not app_id: return None
@@ -72,12 +70,12 @@ def find_game_art(app_id: str | int, platform: str, steam_base: Optional[str]) -
                     art["poster"] = os.path.join(root, t)
                     break
     elif platform == "heroic-epic":
-        paths = download_heroic_assets(app_id, platform)
+        paths = self.download_heroic_assets(app_id, platform)
         return paths.get("art_square") if paths else None
     elif platform == "heroic-gog":
         if isinstance(app_id, list):
             app_id = app_id[0]
-        paths = download_heroic_assets(app_id, platform)
+        paths = self.download_heroic_assets(app_id, platform)
         return paths.get("art_square") if paths else None
     return art
 
@@ -220,3 +218,86 @@ def scan_all_games(game_configs_dir):
             print(f"Error processing {filename} during scan: {e}")
 
     return matches
+
+    # 100% identical
+def download_heroic_assets(appName: str, platform: str):
+    if isinstance(appName, list):
+        appName = str(appName[0])
+    else:
+        appName = str(appName)
+
+    json_path = os.path.expanduser("~/.var/app/com.heroicgameslauncher.hgl/config/heroic/store/download-manager.json") # flatpak
+    if not os.path.exists(json_path):
+        json_path = os.path.expanduser("~/.config/heroic/store/download-manager.json") # not flatpak
+
+    if isinstance(appName, list):
+        appName = appName[0]
+    
+    cache_base = os.path.join(GLib.get_user_data_dir(), "nomm", "image-cache", f"{platform}", f"{appName}")
+    
+    if os.path.exists(cache_base):
+        existing_files = {}
+        for entry in os.listdir(cache_base):
+            if entry.startswith("art_square"):
+                existing_files["art_square"] = os.path.join(cache_base, entry)
+            elif entry.startswith("art_hero"):
+                existing_files["art_hero"] = os.path.join(cache_base, entry)
+        
+        if "art_square" in existing_files:
+            print(f"Using cached assets for {appName}")
+            return existing_files
+
+    if not os.path.exists(json_path):
+        print(f"Heroic config not found at {json_path}")
+        return None
+
+    try:
+        with open(json_path, 'r') as f:
+            data = json.load(f)
+        
+        finished_apps = data.get("finished", [])
+        target_info = None
+
+        for entry in finished_apps:
+            params = entry.get("params", {})
+            game_info = params.get("gameInfo", {})
+            
+            if params.get("appName") == appName or game_info.get("title") == appName:
+                target_info = game_info
+                break
+        
+        if not target_info:
+            return None
+
+        urls = {
+            "art_square": target_info.get("art_square"),
+            "art_hero": target_info.get("art_background") or target_info.get("art_cover")
+        }
+
+        os.makedirs(cache_base, exist_ok=True)
+        downloaded_paths = {}
+
+        for key, url in urls.items():
+            if not url:
+                continue
+                
+            ext = os.path.splitext(url)[1] if "." in url.split("/")[-1] else ".jpg"
+            if "?" in ext: ext = ext.split("?")[0]
+            
+            local_path = os.path.join(cache_base, f"{key}{ext}")
+
+            try:
+                r = requests.get(url, timeout=15)
+                if r.status_code == 200:
+                    with open(local_path, 'wb') as f:
+                        f.write(r.content)
+                    downloaded_paths[key] = local_path
+                    print(f"Downloaded: {local_path}")
+            except Exception as e:
+                print(f"Error downloading {key}: {e}")
+
+        return downloaded_paths
+
+    except Exception as e:
+        print(f"Failed to process Heroic JSON: {e}")
+        return None
